+++ pcweb/components/navbar/__init__.py
@@ -1,0 +1,1 @@
+from .navbar import *
