+++ pcweb/components/navbar/buttons/style.py
@@ -1,0 +1,11 @@
+import reflex as rx
+
+button_style = {
+    "border_radius": "50px",
+    "border": f"1px solid {rx.color('mauve', 4)}",
+    "background": f"1px solid {rx.color('mauve', 2)}",
+    "box_shadow": "0px 3px 7px -4px rgba(21, 18, 44, 0.15)",
+    "padding": "7px 12px 7px 8px",
+    "align_items": "center",
+}
+
