+++ pcweb/components/sidebar/__init__.py
@@ -1,0 +1,1 @@
+from .sidebar import *
