+++ pcweb/components/sidebar/style.py
@@ -1,0 +1,20 @@
+from pcweb import styles
+from pcweb.styles import font_weights as fw
+
+heading_style2 = {
+    "font_size": styles.TEXT_FONT_SIZE,
+    "color": "#5646ED",
+    "background_color": "#F5EFFE",
+    "border_radius": "0.5em",
+    "font_weight": "500",
+    "width": "100%",
+    "padding_x": "0.5em",
+    "padding_y": "0.25em",
+}
+heading_style3 = {
+    "font_weight": fw["section"],
+    "font_size": styles.TEXT_FONT_SIZE,
+    "color": "#696287",
+    "margin_bottom": "0.5em",
+    "margin_left": "1.1em",
+}
